Calculate fee amount denominated in fee token decimals
