+++ scripts/local-deployment/localDeploymentLib.ts
@@ -1,4 +1,4 @@
-import { ethers } from 'ethers'
+import { BigNumber, ethers } from 'ethers'
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { L1Network, L2Network, addCustomNetwork } from '@arbitrum/sdk'
 import { Bridge__factory } from '@arbitrum/sdk/dist/lib/abi/factories/Bridge__factory'
@@ -11,6 +11,7 @@ import {
   getEstimateForDeployingFactory,
 } from '../atomicTokenBridgeDeployer'
 import { l2Networks } from '@arbitrum/sdk/dist/lib/dataEntities/networks'
+import { ERC20__factory } from '../../build/types'
 
 const LOCALHOST_L2_RPC = 'http://localhost:8547'
 const LOCALHOST_L3_RPC = 'http://localhost:3347'
@@ -256,5 +257,28 @@ export const getLocalNetworks = async (
   return {
     l1Network,
     l2Network,
+  }
+}
+
+/**
+ * Scale the amount from 18-denomination to the fee token decimals denomination
+ */
+export async function _getScaledAmount(
+  feeToken: string,
+  amount: BigNumber,
+  provider: ethers.providers.Provider
+): Promise<BigNumber> {
+  const decimals = await ERC20__factory.connect(feeToken, provider).decimals()
+  if (decimals == 18) {
+    return amount
+  } else if (decimals < 18) {
+    let scaledAmount = amount.div(BigNumber.from(10).pow(18 - decimals))
+    // round up if necessary
+    if (scaledAmount.mul(BigNumber.from(10).pow(18 - decimals)).lt(amount)) {
+      scaledAmount = scaledAmount.add(1)
+    }
+    return scaledAmount
+  } else {
+    return amount.mul(BigNumber.from(10).pow(decimals - 18))
   }
 }
